Update ion-hash-go to use Symbol Tokens (#62)
* Update hash to use Symbol Tokens

* updates

* more updates

* More updates

* More updates

* Define AnnotationAsSymbol()

* Define SymbolTokenString()

* Update

* Fix issues

* Update ion_hash_test.go

* Update ion-go dependency

* Update ion_hash_test.go

* Fix TODO

+++ err.go
@@ -57,7 +57,7 @@ func (e *InvalidIonTypeError) Error() string {
 
 // UnknownSymbolError is returned when processing an unknown field name symbol.
 type UnknownSymbolError struct {
-	sid int
+	sid int64
 }
 
 func (e *UnknownSymbolError) Error() string {
